requirements.txt
# RobustLMP-GAN — pinned dependencies
# Python 3.10+

# Core data
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0        # for reading EIA .xlsx files
xlrd>=2.0.1            # for reading legacy .xls files

# Machine learning
torch>=2.1.0
scikit-learn>=1.3.0

# API access
requests>=2.31.0

# Visualisation (optional but recommended)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Experiment tracking (optional)
mlflow>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
